// frontend/src/types/agent.ts
  export interface Agent {
    _id: string;
    _key: string;
    name: string;
    description: string;
    startMessage: string;
    systemPrompt: string;
    tools: string[];
    models: {
      provider: string;
      modelName: string;
      isReasoning: boolean;
    }[];
    apps: string[];
    kb: string[];
    vectorDBs: string[];
    tags: string[];
    templateId?: string;
    createdBy: string;
    orgId: string;
    createdAtTimestamp: string;
    updatedAtTimestamp: string;
    lastUsedAt?: string;
    conversationCount: number;
    sharedWith: string[];
    version: number;
    icon?: string;
    flow?: {
      nodes: any[];
      edges: any[];
    };
  }